Many. MANY bugfixes with regard to process forking

+++ phpversion/Program.class.php
@@ -4,29 +4,53 @@ class Process {
     
 	public $_attribStat = array(
         "name" => false,
-        "pid" => NULL,
+        "pid" => 0,
         "lcmd" => NULL,
-        "pcount" => NULL,
-        "autostart" => false,
-        "rstart_cond" => NULL,
-        "exitcodes" => 0,
-        "startwait" => NULL,
-        "retry" => false,
-        "exitsig" => NULL,
-        "killwait" => NULL,
-        "pid_logging" => false,
-        "pid_logfile" => NULL,
-        "env_vars" => NULL,
+        "pcount" => 2,
+        "autostart" => FALSE,
+        "rstart_cond" => FALSE,
+        "exitcodes" => array(0),
+        "startwait" => 5,
+        "retry" => 3,
+        "exitsig" => 3,
+        "killwait" => 0,
+        "pid_logging" => TRUE,
+        "pid_logfile" => "ls.txt",
+        "env_vars" => array(),
         "wrk_dir" => NULL,
         "umask" => NULL,
-        "status" => false,
-        "restartMe" => false,
+        "status" => FALSE,
+        "restartMe" => FALSE,
         "child" => NULL,
 		"stream" => NULL,
-		"aborted" => FALSE
+		"aborted" => FALSE,
+		"reported" => FALSE
     );
 
-    public function __construct($kwargs) {
+	public function __construct($kwargs) {
+        echo "constructor".PHP_EOL;
+        $this->_attribStat['name'] = "ls";
+        $this->_attribStat['pid'] = 0;
+        $this->_attribStat['lcmd'] = "ls";
+        $this->_attribStat['pcount'] = 1;
+        $this->_attribStat['autostart'] = FALSE;
+        $this->_attribStat['rstart_cond'] = FALSE;
+        $this->_attribStat['exitcodes'] = array(0);
+        $this->_attribStat['startwait'] = 0;
+        $this->_attribStat['retry'] = 3;
+        $this->_attribStat['exitsig'] = 3;
+        $this->_attribStat['killwait'] = 0;
+        $this->_attribStat['pid_logging'] = FALSE;
+        $this->_attribStat['pid_logfile'] = NULL;
+        $this->_attribStat['env_vars'] = array();
+        $this->_attribStat['wrk_dir'] = NULL;
+        $this->_attribStat['umask'] = NULL;
+        $this->_attribStat['status'] = FALSE;
+        $this->_attribStat['restartMe'] = FALSE;
+        $this->_attribStat['child'] = NULL;
+    }
+
+    /*public function __construct($kwargs) {
         echo "constructor".PHP_EOL;
         $this->_attribStat['name'] = trim($kwargs['name:'],";");
         $this->_attribStat['pid'] = trim($kwargs['pid:'],";");
@@ -47,7 +71,7 @@ class Process {
         $this->_attribStat['status'] = trim($kwargs['status:'],";");
         $this->_attribStat['restartMe'] = trim($kwargs['restartMe:'],";");
         $this->_attribStat['child'] = trim($kwargs['child:'],";");
-    }
+    }*/
 
 	public function adjust($pargs) {
         echo "pargs: ".$pargs.PHP_EOL;
@@ -70,8 +94,6 @@ class Process {
         echo "destructor".PHP_EOL;
     }
 
-
-
 	/* We should probably consider breaking down the functionality of this method into smaller functions/methods. It's a bit frightening to look at. */
     public function start() 
 	{
@@ -89,10 +111,21 @@ class Process {
 					/* Generate a file descriptor to redirect the programs stdout if the user requested it */
 					if ($this->_attribStat['pid_logging'] == TRUE)
 					{
+						$descriptorspec = array(
+							0 => array("pipe", "r"),  // stdin is a pipe that the child will read from
+							1 => array("file", $this->_attribStat['name'] . $this->_attribStat['pid'] . "txt", "w"),  // stdout is a pipe that the child will write to
+							2 => array("file", "/dev/null", "w")); // stderr is a file to write to
+
 						if ($this->_attribStat['pid_logfile'] != NULL)
-							$descriptorspec['0'] = fopen($this->_attribStat['pid_logfile'], 'a');
-						else
-							$descriptorspec['0'] = fopen($this->_attribStat['name'] . $this->_attribStat['pid'] . "txt", 'a');
+							$descriptorspec[0] = fopen($this->_attribStat['pid_logfile'], 'a');
+					}
+					else
+					{
+						$descriptorspec = array(
+							0 => array("pipe", "r"),  // stdin is a pipe that the child will read from
+							1 => array("file","/dev/null", "w"),  // stdout is a pipe that the child will write to
+							2 => array("file","/dev/null", "w")); // stderr is a file to write to
+
 					}
 					/* assembly loop to retry program initialization until it runs or must be aborted due to too many attempts*/
 					$attempts = 0;
@@ -102,33 +135,42 @@ class Process {
 					if ($this->_attribStat['umask'] != NULL)
 						$mask = umask($this->_attribStat['umask']);
 					/* Create the Process stream and save the pid */
-					$this->_attribStat['stream'] = proc_open("exec " . $this->_attribStat['lcmd'], $descriptorspec, $pipes, $this->_attribStat['wrk_dir'], $env);
-					$proc_details = proc_get_status($this->_attribStat['stream']);
-					$this->_attribStat['pid'] = $proc_details['pid'];
+					$this->_attribStat['stream'] = proc_open("exec " . $this->_attribStat['lcmd'], $descriptorspec, $pipes,  $this->_attribStat['wrk_dir'], $env);
+					if ($this->_attribStat['stream'] != FALSE)
+					{
+						$proc_details = proc_get_status($this->_attribStat['stream']);
+						$this->_attribStat['pid'] = $proc_details['pid'];
+					}
 					/* Set back the umask post fork */
 					if ($this->_attribStat['umask'] != NULL)
 						umask($mask);
 					/* Calculate how much time in seconds to give the program to start before it is considered properly started */
 					$wait = time() + $this->_attribStat['startwait'];
 					/* Continually check for the process status to make sure it's okay, if not, retry until limit is reached, the output abort debug message and close stream */
-					while (time() <= $this->_attribStat['startwait'])
-						if (pcntl_waitpid($proc_details['pid'], $status, WNOHANG) == $this->_attribStat['pid'])
+					$check = 0;
+					while (time() <= $this->_attribStat['startwait'] || $check == 0)
+					{
+						if ($this->_attribStat['stream'] == FALSE || pcntl_waitpid($proc_details['pid'], $status, WNOHANG) == $this->_attribStat['pid'])
 						{
 							if ($attempts <= $this->_attribStat['retry'])
 							{
-								log_message("Program {$this->_attribStat['name']} at process {$this->_attribStat['pid']} failed to start. Retrying for the {$attempts} time.\n");
-								jump(pexec);
+								log_message("Program {$this->_attribStat['name']} at process {$this->_attribStat['pid']} failed to start. Retrying for the {$attempts} time.");
+								goto pexec;
 							}
 							else
 							{
-								log_message("Program {$this->_attribStat['name']} at process {$this->_attribStat['pid']} failed to start after {$this->_attribStat['retry']} attempts\n");
+								log_message("Program {$this->_attribStat['name']} at process {$this->_attribStat['pid']} failed to start after {$this->_attribStat['retry']} attempts");
 								$this->_attribStat['aborted'] = TRUE;
-								proc_close($this->_attribStat['stream']);
+								if ($this->_attribStat['stream'] != FALSE)
+									proc_close($this->_attribStat['stream']);
 								$this->_attribStat['pid'] = 0;
 							}
 						}
+						$check = 1;
+					}
 					/* If succesful, set status to TRUE for online */
 					$this->_attribStat['status'] = TRUE;
+					$this->_attribStat['reported'] = FALSE;
 				}
 				else
 				{
@@ -139,7 +181,7 @@ class Process {
 							$this->_attribStat['child']->start();
 						else
 						{
-							log_message("Program {$this->_attribStat['name']} is creating a child process\n");
+							log_message("Program {$this->_attribStat['name']} is creating a child process");
 							$this->_attribStat['child'] = clone $this;
 							$this->_attribStat['child']->_pcount = $this->_attribStat['pcount'] - 1;
 							$this->_attribStat['child']->start();
@@ -148,10 +190,10 @@ class Process {
 				}
 			}
 			else
-				log_message("Process ID {$this->_attribStat['pid']} already online during start() call\n");
+				log_message("Process ID {$this->_attribStat['pid']} already online during start() call");
 			if ($this->_attribStat['pcount'] > 1)
 			{
-				log_message("Program {$this->_attribStat['name']} is creating a child process\n");
+				log_message("Program {$this->_attribStat['name']} is creating a child process");
 				$this->_attribStat['child'] = clone $this;
 				$this->_attribStat['child']->_pcount = $this->_attribStat['pcount'] - 1;
 				$this->_attribStat['child']->start();
@@ -166,10 +208,10 @@ class Process {
 		$proc_details = proc_get_status($this->_attribStat['stream']);
 		if ($proc_details['running'] == FALSE)
 		{
-			log_message("{$this->_attribStat['name']} {$this->_attribStat['pid']} has reported as 'OFFLINE', due to exit code {$proc_details['exitcode']} \n");
+			log_message("{$this->_attribStat['name']} {$this->_attribStat['pid']} has reported as 'OFFLINE', due to exit code {$proc_details['exitcode']}");
 			if ($this->_attribStat['rstart_cond'] == TRUE && in_array($proc_details['exitcode'], $this->_attribStat['exitcodes']) == FALSE)
 			{
-				log_message("{$this->_attribStat['name']} {$this->_attribStat['pid']} is set to be restarted after recieving an unnaccepted exitcode. Attempting Restart.\n");
+				log_message("{$this->_attribStat['name']} {$this->_attribStat['pid']} is set to be restarted after recieving an unnaccepted exitcode. Attempting Restart.");
 				$this->start();
 			}
 		}
@@ -193,9 +235,9 @@ class Process {
 		else
 		{
 			if ($this->_attribStat['status'] == TRUE)
-				echo "TM > " . $this->_attribStat['name'] . " " . $this->_attribStat['pid'] . " status: \x1b[34mONLINE\x1b[0m\n";
+				echo "<taskmaster/> " . $this->_attribStat['name'] . " " . $this->_attribStat['pid'] . " status: \x1b[34mONLINE\x1b[0m\n";
 			else
-				echo "TM > " . $this->_attribStat['name'] . " " . $this->_attribStat['pid'] . " status: \x1b[31mOFFLINE\x1b[0m\n";
+				echo "<taskmaster/> " . $this->_attribStat['name'] . " " . $this->_attribStat['pid'] . " status: \x1b[31mOFFLINE\x1b[0m\n";
 			if ($this->_attribStat['child'] != NULL)
 				$this->_attribStat['child']->status(TRUE);
 		}
@@ -215,9 +257,11 @@ class Process {
 
 	public function shutdown()
 	{
+		$proc_details = proc_get_status($this->_attribStat['stream']);
 		if ($this->_attribStat['status'] == TRUE)
 		{
-			proc_terminate($this->_attribStat['stream'], $this->_attribStat['exitsig']);
+			if ($proc_details['running'] == TRUE)
+				proc_terminate($this->_attribStat['stream'], $this->_attribStat['exitsig']);
 			sleep($this->_attribStat['killwait']);
 			$this->kill();
 		}
@@ -227,9 +271,14 @@ class Process {
 
 	public function kill()
 	{
+		$proc_details = proc_get_status($this->_attribStat['stream']);
 		if ($this->_attribStat['status'] == TRUE)
 		{
-			proc_terminate($this->_attribStat['stream']);
+			if ($proc_details['running'] == TRUE)
+			{
+				proc_terminate($this->_attribStat['stream']);
+			}
+			proc_close($this->_attribStat['stream']);
 			$this->_attribStat['pid'] = 0;
 			$this->_attribStat['status'] = FALSE;	
 		}
@@ -239,10 +288,14 @@ class Process {
 
 	private function _gen_env()
 	{
-		$env = phpinfo(INFO_ENVIRONMENT);
-		for ($x; $env[$x] != NULL; $x++)
+		ob_start();
+		phpinfo(INFO_ENVIRONMENT);
+		$env = ob_get_contents();
+		ob_end_clean();
+		$env = explode("\n", $env);
+		for ($x = 0; isset($env[$x]) == TRUE; $x++)
 		{
-			for ($y; $this->_attribStat['env_vars'][$y] != NULL; $y++)
+			for ($y = 0; isset($this->_attribStat['env_vars'][$y]) == TRUE; $y++)
 			{
 				$env_pair = explode("=", $env[$x]);
 				$_env_pair = explode("=", $this->_attribStat['env_vars'][$y]);
@@ -250,10 +303,10 @@ class Process {
 					$env[$x] = $this->_attribStat['env_vars'][$y];
 			}
 		}
-		for ($x; $this->_attribStat['env_vars'][$x] != NULL; $x++)
+		for ($x = 0; isset($this->_attribStat['env_vars'][$x]) == TRUE; $x++)
 		{
 			$exists = FALSE;
-			for ($y; $env[$y] != NULL; $y++)
+			for ($y = 0; isset($env[$y]) == TRUE; $y++)
 			{
 				$env_pair = explode("=", $env[$y]);
 				$_env_pair = explode("=", $this->_attribStat['env_vars'][$x]);
@@ -273,10 +326,11 @@ class Process {
 			{
 				$this->_attribStat['child']->shutdown();
 				$this->_attribStat['child']->remove_children();
-		}
+				unset($this->_attribStat['child']);
+			}
 		else
 		{
-			log_message("Program {$this->_attribStat['name']} is creating a child process\n");
+			log_message("Program {$this->_attribStat['name']} is creating a child process");
 			$this->_attribStat['child'] = clone $this;
 			$this->_attribStat['child']->_pcount = $this->_attribStat['pcount'] - 1;
 			$this->_attribStat['child']->reset_pcount();
@@ -288,7 +342,6 @@ class Process {
 		while ($this->_attribStat['child'] != NULL)
 			$this->_attribStat['child']->remove_children();
 	}
-
 }
 
 ?>
